drop the protocol if it's http on start URL
If the URL is http or https, just replace it with gopher. This comes up
due to social media adding the protocol in front of domains without
telling anyone.

Also, URLs that copher doesn't know how to deal with will open in their
default programs. This currently means a blank copher window will also
open.